# 01/spec/api.yml
openapi: 3.0.3
info:
  title: 力麗飯店 LINE OA CRM API
  description: 多渠道會員管理與行銷平台 API，整合 LINE、Facebook Messenger 和 Webchat。
  version: 0.6.0
servers:
  - url: /api/v1
    description: API v1

# ============================================================================
# 既有功能 API（根據現有代碼與 Feature 規格）
# ============================================================================
paths:
...
  /faq/categories:
    get:
      summary: 查詢 FAQ 大分類列表（含啟用狀態、已發布規則數）
      operationId: listFaqCategories
      tags: [FAQ 管理]
      responses:
        '200':
          description: 大分類列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaqCategoryListResponse'
        '403':
          description: 未擁有 faq.view 權限
  /faq/categories/{category_id}/toggle:
    patch:
      summary: 啟用或關閉 FAQ 大分類
      operationId: toggleFaqCategory
      tags: [FAQ 管理]
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_active]
              properties:
                is_active:
                  type: boolean
      responses:
        '200':
          description: 切換成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: 未擁有 faq.manage 權限

  # --------------------------------------------------------------------------
  # FAQ 規則管理
  # Feature: faq_rule_management
  # --------------------------------------------------------------------------
  /faq/categories/{category_id}/rules:
    get:
      summary: 查詢大分類下的規則清單
      operationId: listFaqRules
      tags: [FAQ 管理]
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 規則清單
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaqRuleListResponse'
        '403':
          description: 未擁有 faq.view 權限
    post:
      summary: 在大分類下新增規則（狀態為測試未發佈）
      operationId: createFaqRule
      tags: [FAQ 管理]
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaqRuleCreate'
      responses:
        '200':
          description: 建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaqRuleDetail'
        '400':
          description: 必填欄位未填 或 已達規則上限（20 筆）
        '403':
          description: 未擁有 faq.manage 權限

  # /faq/categories/{category_id}/fields: REMOVED — fields 已內嵌在 category 回應中
  /faq/categories/{category_id}/rules/export:
    get:
      summary: 匯出大分類下所有規則
      operationId: exportFaqRules
      tags: [FAQ 管理]
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 匯出檔案下載
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          description: 未擁有 faq.manage 權限
  /faq/categories/{category_id}/rules/import:
    post:
      summary: 匯入規則（完全覆蓋該大分類下所有現有規則）
      operationId: importFaqRules
      tags: [FAQ 管理]
      description: |
        匯入後「完全覆蓋」並「刪除」目前該分類下的所有規則，此動作無法還原。
        前端須在呼叫前彈出二次確認視窗。
        支援格式：.csv, .xls, .xlsx。
        匯入成功後前端應自動重新整理（Reload）並顯示「規則匯入成功」。
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 規則檔案（僅接受 .csv, .xls, .xlsx）
      responses:
        '200':
          description: 匯入成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported_count:
                    type: integer
                    description: 匯入的規則數量
        '400':
          description: 檔案格式不符或資料驗證失敗
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "檔案格式不符，僅支援 .csv, .xls, .xlsx"
        '403':
          description: 未擁有 faq.manage 權限
  /faq/rules/{rule_id}:
    get:
      summary: 取得單一規則詳細資料（含版本資訊）
      operationId: getFaqRule
      tags: [FAQ 管理]
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 規則詳細資料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaqRuleDetail'
        '404':
          description: 規則不存在
    put:
      summary: 編輯規則內容（狀態變為測試未發佈）
      operationId: updateFaqRule
      tags: [FAQ 管理]
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaqRuleUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaqRuleDetail'
        '400':
          description: 必填欄位未填
        '403':
          description: 未擁有 faq.manage 權限
    delete:
      summary: 刪除規則
      operationId: deleteFaqRule
      tags: [FAQ 管理]
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 刪除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: 未擁有 faq.manage 權限
  /faq/rules/{rule_id}/toggle:
    patch:
      summary: 啟用或停用單一規則
      description: |
        停用規則（action=disable）→ status 變為 disabled。
        重新啟用（action=enable）→ status 變為 draft（測試未發佈），
        須透過 POST /faq/publish 發佈後才生效於前台。
      operationId: toggleFaqRule
      tags: [FAQ 管理]
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [enable, disable]
                  description: enable → draft；disable → disabled
      responses:
        '200':
          description: 切換成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  rule_id:
                    type: integer
                  status:
                    type: string
                    enum: [draft, disabled]
                    description: enable 後為 draft，disable 後為 disabled
        '403':
          description: 未擁有 faq.manage 權限
  /faq/rules/{rule_id}/publish:
    post:
      summary: 發佈單一規則
      operationId: publishFaqRule
      tags: [FAQ 管理]
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 發佈成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaqRuleDetail'
        '404':
          description: 規則不存在
  /faq/rules/{rule_id}/versions:
    get:
      summary: 取得規則版本歷史
      operationId: getFaqRuleVersions
      tags: [FAQ 管理]
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 版本清單
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: array
                    items:
                      type: object
  /faq/rules/{rule_id}/revert:
    post:
      summary: 將規則恢復為上一個已啟用版本
      operationId: revertFaqRule
      tags: [FAQ 管理]
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 恢復成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaqRuleDetail'
        '400':
          description: 無可回復的版本
        '403':
          description: 未擁有 faq.manage 權限
  /faq/publish:
    post:
      summary: 發佈所有測試未發佈狀態的規則至正式環境（即時生效，含會員聊天室熱更新）
      operationId: publishFaqRules
      tags: [FAQ 管理]
      description: |
        發佈後規則立即同步至前台聊天機器人及系統內部 AI Chatbot 會員聊天室。
        會員聊天室若有進行中對話，AI 下一次回覆即引用最新規則，
        並在聊天室顯示「規則已更新」提示。
        前端收到 200 後顯示 Toast「發佈成功」。
      responses:
        '200':
          description: 發佈成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  published_count:
                    type: integer
                    description: 發佈的規則數量
        '403':
          description: 未擁有 faq.publish 權限

  # --------------------------------------------------------------------------
  # FAQ Token 用量
  # Feature: faq_category_management (Token 用量管理)
  # --------------------------------------------------------------------------
...
  /faq/test-chat:
    post:
      summary: 在測試聊天視窗發送提問並取得 AI 回答
      operationId: faqTestChat
      tags: [FAQ 管理]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  description: 使用者提問內容
      responses:
        '200':
          description: AI 回答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaqTestChatResponse'
        '403':
          description: 未擁有 faq.manage 權限

  # --------------------------------------------------------------------------
  # AI 聊天機器人
  # Feature: ai_chatbot_integration
  # --------------------------------------------------------------------------
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          default: 200
        message:
          type: string
          default: success
        data:
          type: object
        timestamp:
          type: string
          format: date-time
...
    FaqCategoryListResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              category_name:
                type: string
              is_active:
                type: boolean
              data_source_type:
                type: string
                enum: [pms, custom_faq]
                description: 資料來源類型（PMS 串接 / 自訂 FAQ）
              source_status:
                type: string
                enum: [not_connected, enabled, disabled]
                description: |
                  來源狀態。PMS 串接有 3 種：未串接(not_connected)、已啟用(enabled)、已停用(disabled)。
                  自訂 FAQ 有 2 種：已啟用(enabled)、已停用(disabled)。
              published_rule_count:
                type: integer
              total_rule_count:
                type: integer
                description: 該大分類下所有規則（含已啟用、已停用、測試未發佈）的總數
              sort_order:
                type: integer
        total_count:
          type: integer
          description: 大分類總數（顯示為「共 N 類」）
        active_count:
          type: integer
          description: 已啟用大分類數（顯示為「已啟用 N 類」）
        last_updated_at:
          type: string
          format: date-time
          description: 全域最後更新時間（觸發條件：新增/編輯/刪除規則、啟停大分類/規則、串接/取消串接 PMS）
...